[CLN] Move data to core
