+++ liquid_earth_api/core/data/schemas.py
